Extract to GemsResolver
